task 14

+++ 14-binary_tree_balance.c
@@ -16,6 +16,6 @@ int binary_tree_balance(const binary_tree_t *tree)
 	if (!tree)
 		return (0);
 
-	return (binary_tree_height(tree->left)
-			- binary_tree_height(tree->right));
+	return ((int)(binary_tree_height(tree->left)
+			- binary_tree_height(tree->right)));
 }
